requirements.txt
Django==1.8.7
djangorestframework==3.3.1
requests==0.13.2
stripe==1.19.1
wheel==0.24.0
